Module curio::structs

Structs

Cookie
The structure used for storing Cookies and their configuration

CurioConfig

Header
The structure used for internal parsing of headers

PostData
store request body content through a variety of methods, resulting in similar, or identical outputs via HTTP POST

Request
Build a request, without any of the knowledge of how HTTP works, this structure takes care of it all, and still follows specifications

Response
All the nitty gritty of a Response, put into a nicely formatted, easy to use structure for you to use

Enums

DataType

HTTPVersion
Defines the type of HTTP to be used in the request (TCP/TLS)

RequestType
Defines the method to be used in the request